refactor: move progress reports into directory
